% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGaoDeMap.R
\name{changeZoom}
\alias{changeZoom}
\title{Change the zoom seting of a ggmap object.}
\usage{
changeZoom(map, location, width, height, zoom)
}
\arguments{
\item{map}{ggmap object to change zoom}

\item{location}{a vector a or matrix contains longtitude and latitude of the center of the map.}

\item{width}{width of the map}

\item{height}{height of the map}

\item{zoom}{map zoom, an integer from 3 (continent) to 21 (building), default value 10 (city)}
}
\value{
A ggmap object.
}
\description{
Take in a ggmap object and set it's zoom(ll,ur coordinate).
}
