% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoordinate.R
\name{getCoordinate}
\alias{getCoordinate}
\title{Get coordiante from address. 
Take in address and return the coordinate}
\usage{
getCoordinate(address, city = NULL, output = "json", formatted = F,
  limit = 600, map_ak = "")
}
\arguments{
\item{address}{address}

\item{city}{the city of the address, optional}

\item{output}{should be "json" or "xml", the type of the result}

\item{formatted}{logical value, return the coordinates or the original results}

\item{limit}{integer value.If the length of address exceeded limit, function will run in parallel}

\item{map_ak}{access key of GaoDe map service, you can set it through options(gaode.key='xxx')}
}
\value{
A vector contains the  corresponding coordiante. If "formatted=TRUE", return the numeric coordinates, otherwise return json or xml type result, depents on the argument "output". If the length of address is larger than 1, the result is a matrix.
}
\description{
Get coordiante from address. 
Take in address and return the coordinate
}
\examples{
\dontrun{ 
## json output
getCoordinate('鍖椾含澶у')

## xml output
getCoordinate('鍖椾含澶у', output='xml')

## formatted
getCoordinate('鍖椾含澶у', formatted = T)

## vectorization, return a matrix
getCoordinate(c('鍖椾含澶у', '娓呭崕澶у'), formatted = T)
}
}
