% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGaoDeMap.R
\name{getGaoDeMap}
\alias{getGaoDeMap}
\title{Get static map from GaoDe API}
\usage{
getGaoDeMap(location, city = NULL, width = 400, height = 400,
  zoom = 10, scale = 2, color = "color", messaging = TRUE,
  map_ak = "", zoomplus = 0)
}
\arguments{
\item{location}{a vector a or matrix contains longtitude and latitude of the center of the map, or a character refers to the address}

\item{city}{city of location}

\item{width}{width of the map}

\item{height}{height of the map}

\item{zoom}{map zoom, an integer from 1 (continent) to 17 (building), default value 10 (city)}

\item{scale}{multiplicative factor for the number of pixels returned. possible values are 2 or anything else.}

\item{color}{"color" or "bw", color or black-and-white}

\item{messaging}{logical. whether to print the download messages.}

\item{zoomplus}{map zoom plus, a real number for fine turning zoom in further ggplot commands}

\item{coordtype}{coordiantion type.}
}
\value{
A ggmap object. a map image as a 2d-array of colors as hexadecimal strings representing pixel fill values.
}
\description{
Take in location and return a ggmap object.
}
\examples{

\dontrun{
library(ggmap)  
## Beijing
p <- getGaoDeMap(c(116.39565, 39.92999))
ggmap(p)

p <- getGaoDeMap('beijing') # the same
ggmap(p)

## black-and-white
p <- getGaoDeMap(color='bw')
ggmap(p)

## do not print messages
p <- getGaoDeMap(messaging = F)
}
}
