% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlace.R
\name{getPlace}
\alias{getPlace}
\title{Get place from query
Take in query and city, return the informations}
\usage{
getPlace(place = NULL, city = "北京", page_size = 20,
  pages = Inf, extensions = "base", verbose = TRUE, map_ak = "")
}
\arguments{
\item{place}{the place you want to search}

\item{city}{define the city}
}
\value{
a data frame contains name, longtitude, latitude and address, as well as teleplhone number, if exist.
}
\description{
Get place from query
Take in query and city, return the informations
}
\examples{
\dontrun{
## colleges in beijing
bj_college = getPlace('澶у', '鍖椾含')
## Mcdonald's in shanghai
sh_mcdonald = getPlace('楹﹀綋鍔戼㸳', '涓婃捣')
}
}
