% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegionCoordinates.R
\name{getRegionCoordinates}
\alias{getRegionCoordinates}
\title{Get edge coordinates of a given district.
Take in a district name and return a dataframe of coordinates of the edge of the district.}
\usage{
getRegionCoordinates(target, ak, id = "1")
}
\arguments{
\item{target}{the name of the district.}

\item{id}{the id of the district}
}
\description{
Get edge coordinates of a given district.
Take in a district name and return a dataframe of coordinates of the edge of the district.
}
\details{
Go to https://lbs.amap.com/api/webservice/guide/api/district to read more.
}
