% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRoute.R
\name{getRoute}
\alias{getRoute}
\title{Get route from query
Take in a original location and destination, return the direction}
\usage{
getRoute(origin, destination, mode = "driving", city = NULL,
  strategy = 9, size = 2, output = "xml", map_ak = "")
}
\arguments{
\item{origin}{the original location}

\item{destination}{the destination}

\item{mode}{'driving'(default), 'walking', or 'transit','bicycling','truck'.}

\item{city}{the city of of original location and destination.}

\item{strategy}{integer from 0 to 20 for path selecting, default 9 for driving mode.}

\item{size}{for truck mode.}
}
\value{
a data frame contains longtitude and latitude of the route.
}
\description{
Get route from query
Take in a original location and destination, return the direction
}
\details{
Go to https://lbs.amap.com/api/webservice/guide/api/direction to read more.
}
\examples{
\dontrun{
bjMap = getBaiduMap('鍖椾含', color='bw')
df = getRoute('棣栭兘鍥介檯鏈哄満', '鍖椾含鍗楄嫅鏈哄満', region = '鍖椾含')
ggmap(bjMap) + geom_path(data = df, aes(lon, lat), alpha = 0.5, col = 'red')
}

}
